requirements.txt
pybithumb
matplotlib
websocket-client
slack_sdk